+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ### Changed
 - Indexed all foreign key references to users [#5611](https://github.com/raster-foundry/raster-foundry/pull/5611)
+- Use original filename for COG [#5613](https://github.com/raster-foundry/raster-foundry/pull/5613)
 
 ### Fixed
 - Updated some setup steps and Auth0 interaction for more convenient external user use [#5612](https://github.com/raster-foundry/raster-foundry/pull/5612)